+++ CHANGELOG.md
@@ -10,6 +10,7 @@ ENHANCEMENTS:
 BUG FIXES:
 * Fix circular dependancy in base workspace. ([#4756](https://github.com/microsoft/AzureTRE/pull/4756))
 * Replaced deprecated `datetime.utcnow()` with `datetime.now(datetime.UTC)` in the API and airlock processor. ([#4743](https://github.com/microsoft/AzureTRE/issues/4743))
+* Allow Service Bus Standard deployments to reach Service Bus from the resource processor and API app subnets via firewall service tag rules. ([#4256](https://github.com/microsoft/AzureTRE/pull/4256))
 
 COMPONENTS: